refactor PixelOperationsTests
